requirements.txt
langchain==0.3.7
langchain-community==0.3.7
faiss-cpu==1.7.4
rank-bm25==0.2.2
pymupdf==1.23.8
arxiv==2.1.0
fastapi==0.115.0
uvicorn[standard]==0.30.0
pydantic==2.9.0
pydantic-settings==2.5.0
ollama==0.3.3

sentence-transformers==3.0.1 
huggingface-hub==0.24.0
